requirements.txt
accelerate==0.34.2
adjustText==1.3.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alignn==2023.8.1
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
ase==3.23.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
Brotli @ file:///croot/brotli-split_1714483155106/work
catboost==1.2.8
certifi @ file:///croot/certifi_1725551672989/work/certifi
cffi==1.17.1
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
colour-science==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
datasets==3.1.0
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
dgl==1.0.2+cu118
dill==0.3.8
docker-pycreds==0.4.0
e3nn==0.5.1
entrypoints==0.4
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
filelock @ file:///croot/filelock_1700591183607/work
flake8==7.1.1
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
graphviz==0.20.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.7
idna==3.8
imageio==2.37.0
ipykernel==6.29.5
ipython==8.27.0
ipython-genutils==0.2.0
ipywidgets==8.1.5
isoduration==20.11.0
jarvis-tools==2024.8.30
jedi==0.19.1
Jinja2 @ file:///croot/jinja2_1716993405101/work
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
latexcodec==3.0.0
lightning-utilities==0.11.7
llvmlite==0.44.0
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.7.0
matplotlib-inline==0.1.7
mccabe==0.7.0
-e git+https://github.com/atomisticnet/MeLting@52b3d97bafee1fceb98e0fb6a41d1e1886f7e193#egg=MeLting
mistune==3.0.2
mkl-service==2.4.0
mkl_fft @ file:///croot/mkl_fft_1725370245198/work
mkl_random @ file:///croot/mkl_random_1725370241878/work
monty==2024.7.30
mpmath @ file:///croot/mpmath_1690848262763/work
multidict==6.1.0
multiprocess==0.70.16
nbclassic==1.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx @ file:///croot/networkx_1720002482208/work
notebook==6.5.7
notebook_shim==0.2.4
numba==0.61.2
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp310-cp310-linux_x86_64.whl#sha256=d8cd837ed43e87f77e6efaa08e8de927ca030a1c9c5d04624432d6fb9a74a5ee
openpyxl==3.1.5
opt-einsum==3.3.0
opt-einsum-fx==0.1.4
overrides==7.7.0
p_tqdm==1.4.2
packaging==24.1
palettable==3.3.3
pandarallel==1.6.5
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.4
pathos==0.3.3
patsy==1.0.1
pexpect==4.9.0
pillow @ file:///croot/pillow_1721059439630/work
platformdirs==4.3.2
plotly==5.24.1
pox==0.3.5
ppft==1.7.6.9
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==5.28.3
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pybtex==0.24.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==1.8.1
pydocstyle==6.3.0
pyflakes==3.2.0
Pygments==2.18.0
pymatgen==2024.9.17
pynndescent==0.5.13
pyparsing==2.4.7
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytorch-ignite==0.5.0.dev20230428
pytorch-lightning==2.0.4
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
qtconsole==5.6.0
QtPy==2.4.1
referencing==0.35.1
requests @ file:///croot/requests_1721410876868/work
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.5
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.17.0
setproctitle==1.3.3
shap==0.45.0
six==1.16.0
slicer==0.0.7
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
spglib==2.5.0
stack-data==0.6.3
statsmodels==0.14.4
sympy @ file:///croot/sympy_1724938189289/work
tabulate==0.9.0
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
tomli==2.0.1
toolz==0.12.1
torch==2.1.2
torch-cluster==1.6.3+pt21cu118
torch-geometric==2.6.1
torch-scatter==2.1.2+pt21cu118
torch-sparse==0.6.18+pt21cu118
torchaudio==2.1.2
torchmetrics==1.4.1
torchvision==0.16.2
tornado==6.4.1
tqdm @ file:///croot/tqdm_1724853939799/work
traitlets==5.14.3
triton==2.1.0
types-python-dateutil==2.9.0.20240906
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.1
umap-learn==0.5.7
uncertainties==3.2.2
uri-template==1.3.0
urllib3 @ file:///croot/urllib3_1718912636303/work
wandb==0.18.5
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
xgboost==1.7.4
xmltodict==0.13.0
xxhash==3.5.0
yarl==1.11.1
